[rt.cpan.org 73353] Use ports instead of schemes when resolving hosts.
Solaris (and probably other systems) don't have an /etc/services entry
for http, so relying on the lookup fails here.  Luckily the user also
provides a port (which may or may not be symbolic).  I'll use that for
now and see where it gets me.

Thank you, Michael Lackhoff, for reporting this bug.