[rt.cpan.org 76776] Fix an edge case where queued requests don't start.
Sergei Kozunov came through with another great ticket.  If two
requests for the same host arrive, and the first one fails, the second
one will time out in the queue.  But only if no other requests are
already pending.  This change wakes up the queue when a connection
fails, causing the subsequent request to be processed.